+++ core/terraform/dns_zones_non_core.tf
@@ -3,7 +3,7 @@
 
 resource "azurerm_private_dns_zone" "non_core" {
   for_each            = local.private_dns_zone_names_non_core
-  name                = each.key
+  name                = module.terraform_azurerm_environment_configuration.private_links[each.key]
   resource_group_name = azurerm_resource_group.core.name
   tags                = local.tre_core_tags
 
@@ -18,5 +18,14 @@ resource "azurerm_private_dns_zone_virtual_network_link" "mysql" {
   name                  = azurerm_private_dns_zone.non_core["privatelink.mysql.database.azure.com"].name
   registration_enabled  = false
   tags                  = local.tre_core_tags
+  lifecycle { ignore_changes = [tags] }
+}
+
+# Once the deployment of the app gateway is complete, we can proceed to include the required DNS zone for Nexus, which is dependent on the FQDN of the app gateway.
+resource "azurerm_private_dns_zone" "nexus" {
+  name                = "nexus-${module.appgateway.app_gateway_fqdn}"
+  resource_group_name = azurerm_resource_group.core.name
+  tags                = local.tre_core_tags
+
   lifecycle { ignore_changes = [tags] }
 }