+++ core/terraform/network/locals.tf
@@ -27,6 +27,7 @@ locals {
     tre_core_service_id = var.tre_id
   }
 
+
   private_dns_zone_names = toset([
     "privatelink.queue.core.windows.net",
     "privatelink.table.core.windows.net"