+++ resource_processor/shared/config.py
@@ -17,7 +17,11 @@ def get_config(logger_adapter) -> dict:
     config["service_bus_namespace"] = os.environ["SERVICE_BUS_FULLY_QUALIFIED_NAMESPACE"]
     config["vmss_msi_id"] = os.environ.get("VMSS_MSI_ID", None)
     config["number_processes"] = os.environ.get("NUMBER_PROCESSES", "1")
-    config["key_vault_name"] = os.environ.get("KEY_VAULT_NAME", os.environ.get("KEYVAULT", None))
+    config["key_vault_url"] = os.environ.get("KEY_VAULT_URL", os.environ.get("KEYVAULT", None))
+    config["arm_environment"] = os.environ.get("ARM_ENVIRONMENT", "public")
+    config["azure_environment"] = os.environ.get("AZURE_ENVIRONMENT", "AzureCloud")
+    config["aad_authority_url"] = os.environ.get("AAD_AUTHORITY_URL", "https://login.microsoftonline.com")
+    config["microsoft_graph_fqdn"] = os.environ.get("MICROSOFT_GRAPH_FQDN", "graph.microsoft.com")
 
     try:
         config["number_processes_int"] = int(config["number_processes"])
@@ -45,7 +49,9 @@ def get_config(logger_adapter) -> dict:
     config["porter_env"] = {
         "HOME": os.environ["HOME"],
         "PATH": os.environ["PATH"],
-        "KEY_VAULT_NAME": config["key_vault_name"],
+        "KEY_VAULT_URL": config["key_vault_url"],
+        "ARM_ENVIRONMENT": config["arm_environment"],
+        "AZURE_ENVIRONMENT": config["azure_environment"],
 
         # These are needed since they are referenced as credentials in every bundle and also in arm_auth credential set.
         "ARM_CLIENT_ID": config["arm_client_id"],