+++ templates/workspace_services/health-services/terraform/providers.tf
@@ -30,3 +30,8 @@ provider "azurerm" {
     }
   }
 }
+
+module "terraform_azurerm_environment_configuration" {
+  source          = "git::https://github.com/microsoft/terraform-azurerm-environment-configuration.git?ref=0.2.0"
+  arm_environment = var.arm_environment
+}