+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 A migration for OperationSteps in Operation objects was added ([#3358](https://github.com/microsoft/AzureTRE/pull/3358)).
 
 FEATURES:
+* (UI) Added upgrade button to resources that have pending template upgrades ([#3387](https://github.com/microsoft/AzureTRE/pull/3387))
+* Enable deployment to Azure Government Cloud ([#3128](https://github.com/microsoft/AzureTRE/issues/3128)).
 
 ENHANCEMENTS:
 * Added 'availableUpgrades' field to Resources in GET/GET all Resources endpoints. The field indicates whether there are template versions that a resource can be upgraded to [#3234](https://github.com/microsoft/AzureTRE/pull/3234)